Deleted Base Viewer GUI
It didnt work it just broke the cutlines and i couldnt figure out how to fix.

so we are starting from the start as Matt is using a different library for translation and rotation i wont add any more to it
